fix(tests): move unit tests to a separate file
