+++ src/client/panes/unit/grid_tests.rs
@@ -1,0 +1,85 @@
+use super::super::Grid;
+use ::insta::assert_snapshot;
+
+fn read_fixture(fixture_name: &str) -> Vec<u8> {
+    let mut path_to_file = std::path::PathBuf::new();
+    path_to_file.push("src");
+    path_to_file.push("tests");
+    path_to_file.push("fixtures");
+    path_to_file.push(fixture_name);
+    let content = std::fs::read(path_to_file)
+        .unwrap_or_else(|_| panic!("could not read fixture {:?}", &fixture_name));
+    content
+}
+
+#[test]
+fn vttest1_0() {
+    let mut vte_parser = vte::Parser::new();
+    let mut grid = Grid::new(41, 110);
+    let fixture_name = "vttest1-0";
+    let content = read_fixture(fixture_name);
+    for byte in content {
+        vte_parser.advance(&mut grid, byte);
+    }
+    assert_snapshot!(format!("{:?}", grid));
+}
+
+#[test]
+fn vttest1_1() {
+    let mut vte_parser = vte::Parser::new();
+    let mut grid = Grid::new(41, 110);
+    let fixture_name = "vttest1-1";
+    let content = read_fixture(fixture_name);
+    for byte in content {
+        vte_parser.advance(&mut grid, byte);
+    }
+    assert_snapshot!(format!("{:?}", grid));
+}
+
+#[test]
+fn vttest1_2() {
+    let mut vte_parser = vte::Parser::new();
+    let mut grid = Grid::new(41, 110);
+    let fixture_name = "vttest1-2";
+    let content = read_fixture(fixture_name);
+    for byte in content {
+        vte_parser.advance(&mut grid, byte);
+    }
+    assert_snapshot!(format!("{:?}", grid));
+}
+
+#[test]
+fn vttest1_3() {
+    let mut vte_parser = vte::Parser::new();
+    let mut grid = Grid::new(41, 110);
+    let fixture_name = "vttest1-3";
+    let content = read_fixture(fixture_name);
+    for byte in content {
+        vte_parser.advance(&mut grid, byte);
+    }
+    assert_snapshot!(format!("{:?}", grid));
+}
+
+#[test]
+fn vttest1_4() {
+    let mut vte_parser = vte::Parser::new();
+    let mut grid = Grid::new(41, 110);
+    let fixture_name = "vttest1-4";
+    let content = read_fixture(fixture_name);
+    for byte in content {
+        vte_parser.advance(&mut grid, byte);
+    }
+    assert_snapshot!(format!("{:?}", grid));
+}
+
+#[test]
+fn vttest1_5() {
+    let mut vte_parser = vte::Parser::new();
+    let mut grid = Grid::new(41, 110);
+    let fixture_name = "vttest1-5";
+    let content = read_fixture(fixture_name);
+    for byte in content {
+        vte_parser.advance(&mut grid, byte);
+    }
+    assert_snapshot!(format!("{:?}", grid));
+}
